+++ lib/src/models/customer/custom_field/create_custom_field_response_model.dart
@@ -47,7 +47,7 @@ class MxCreateCustomFieldResponseModel {
       'isSystem': isSystem,
       'isDeleted': isDeleted,
       'isRequired': isRequired,
-      'options': options,
+      'options': options.map((e) => e.toJson()).toList(),
     }..removeWhere((_, v) => v == null);
   }