+++ lib/src/services/customer_custom_field_service_impl.dart
@@ -3,7 +3,8 @@ import 'package:mx_merchant/src/models/customer/custom_field/create_custom_field
 import 'package:mx_merchant/src/models/customer/custom_field/get_custom_field_response_model.dart';
 import 'package:mx_merchant/src/services/api_service.dart';
 import 'package:mx_merchant/src/services/customer_custom_field_service.dart';
-import 'package:mx_merchant/src/utils/mx_merchant_exception.dart';
+
+import '../utils/throw_error.dart';
 
 class MxCustomFieldServiceImpl implements MxCustomFieldService {
   late ApiService _apiService;
@@ -21,39 +22,27 @@ class MxCustomFieldServiceImpl implements MxCustomFieldService {
     if (response.statusCode == 201) {
       return MxCreateCustomFieldResponseModel.fromJson(Map<String, dynamic>.from(response.data));
     } else {
-      throw MxMerchantException(
-        statusCode: response.statusCode ?? 500,
-        message: response.data['message']?.toString() ?? 'MX Merchant API error',
-        response: response.data,
-      );
+      throw errorParser(response);
     }
   }
 
   @override
   Future<List<MxGetCustomFieldResponseModel>> get(String customerId) async {
-    final response = await _apiService.post('$_route/customer/field', query: {'id': customerId});
+    final response = await _apiService.get('$_route/customer/field', query: {'id': customerId});
     if (response.statusCode == 200) {
       return MxGetCustomFieldResponseModel.fromJsonArray(response.data);
     } else {
-      throw MxMerchantException(
-        statusCode: response.statusCode ?? 500,
-        message: response.data['message']?.toString() ?? 'MX Merchant API error',
-        response: response.data,
-      );
+      throw errorParser(response);
     }
   }
 
   @override
   Future<bool> delete(int fieldId) async {
     final response = await _apiService.delete('$_route/field/$fieldId');
-    if (response.statusCode == 200) {
+    if (response.statusCode == 204) {
       return true;
     } else {
-      throw MxMerchantException(
-        statusCode: response.statusCode ?? 500,
-        message: response.data['message']?.toString() ?? 'MX Merchant API error',
-        response: response.data,
-      );
+      throw errorParser(response);
     }
   }
 }